Moved notifications to it's own tab
