+++ admin/src/page/Settings.js
@@ -5,11 +5,13 @@ import { withRouter, Link, Switch, Route } from "react-router-dom";
 import { ReactComponent as GeneralIcon } from "../assets/svg/settings-general.svg";
 import { ReactComponent as ConsoleIcon } from "../assets/svg/console.svg";
 import { ReactComponent as FilterIcon } from "../assets/svg/filter.svg";
+import { ReactComponent as NotificationsIcon } from "../assets/svg/notifications.svg";
 import Console from "./settings/console";
 import General from "./settings/general";
 import Radarr from "./settings/radarr";
 import Sonarr from "./settings/sonarr";
 import Filter from "./settings/filter";
+import Notifications from "./settings/notifications";
 
 class Settings extends React.Component {
   constructor(props) {
@@ -87,6 +89,18 @@ class Settings extends React.Component {
             </div>
           </Link>
           <Link
+            to="/settings/notifications"
+            className={
+              "settings--menu--item " +
+              (current === "/settings/notifications" ? "active" : "")
+            }
+          >
+            <p>Notifications</p>
+            <div className="icon">
+              <NotificationsIcon />
+            </div>
+          </Link>
+          <Link
             to="/settings/console"
             className={
               "settings--menu--item " +
@@ -110,11 +124,14 @@ class Settings extends React.Component {
             <Route path="/settings/sonarr">
               <Sonarr msg={this.props.msg} />
             </Route>
-            <Route path="/settings/console">
-              <Console msg={this.props.msg} />
-            </Route>
             <Route path="/settings/filter">
               <Filter msg={this.props.msg} />
+            </Route>
+            <Route path="/settings/notifications">
+              <Notifications msg={this.props.msg} />
+            </Route>
+            <Route path="/settings/console">
+              <Console msg={this.props.msg} />
             </Route>
             <Route path="*" exact>
               <section>
